+++ rs/nns/governance/api/src/ic_nns_governance.pb.v1.rs
@@ -652,13 +652,13 @@ pub mod proposal {
         /// Create a new SNS.
         #[prost(message, tag = "24")]
         CreateServiceNervousSystem(super::CreateServiceNervousSystem),
-        /// Calls install_code for a canister (install, reinstall, or upgrade).
+        /// Install, reinstall or upgrade the code of a canister that is controlled by the NNS.
         #[prost(message, tag = "25")]
         InstallCode(super::InstallCode),
-        /// Stops or starts a canister controlled by Root.
+        /// Stop or start a canister that is controlled by the NNS.
         #[prost(message, tag = "26")]
         StopOrStartCanister(super::StopOrStartCanister),
-        /// Updates canister settings for those controlled by NNS Root.
+        /// Update the settings of a canister that is controlled by the NNS.
         #[prost(message, tag = "27")]
         UpdateCanisterSettings(super::UpdateCanisterSettings),
     }
@@ -1548,11 +1548,6 @@ pub struct ProposalData {
     /// This is populated when an OpenSnsTokenSwap proposal is first made.
     #[prost(uint64, optional, tag = "17")]
     pub original_total_community_fund_maturity_e8s_equivalent: Option<u64>,
-    /// This is populated when OpenSnsTokenSwap is executed. It is used when our
-    /// conclude_community_fund_participation Candid method is called to either
-    /// mint ICP, or restore CF neuron maturity.
-    #[prost(message, repeated, tag = "18")]
-    pub cf_participants: Vec<::ic_sns_swap::pb::v1::CfParticipant>,
     /// This gets set to one of the terminal values (i.e. Committed or Aborted)
     /// when the swap canister calls our conclude_community_fund_participation
     /// Candid method. Initially, it is set to Open, because swap is supposed to
@@ -3879,8 +3874,8 @@ pub enum Topic {
     /// creating new subnets, adding and removing subnet nodes, and
     /// splitting subnets.
     SubnetManagement = 7,
-    /// Installing and upgrading “system” canisters that belong to the network.
-    /// For example, upgrading the NNS.
+    /// All proposals to manage NNS-controlled canisters not covered by other topics (Protocol
+    /// Canister Management or Service Nervous System Management).
     NetworkCanisterManagement = 8,
     /// Proposals that update KYC information for regulatory purposes,
     /// for example during the initial Genesis distribution of ICP in the
@@ -3916,11 +3911,11 @@ pub enum Topic {
     ApiBoundaryNodeManagement = 15,
     /// Proposals related to subnet rental.
     SubnetRental = 16,
-    /// Proposals to manage protocol canisters. Those are canisters that are considered part of the IC
-    /// protocol, without which the IC will not be able to function properly.
+    /// All proposals to manage protocol canisters, which are considered part of the ICP protocol
+    /// and are essential for its proper functioning.
     ProtocolCanisterManagement = 17,
-    /// Proposals related to Service Nervous System (SNS) - (1) upgrading SNS-W, (2) upgrading SNS
-    /// Aggregator, and (3) adding WASM's or custom upgrade paths to SNS-W.
+    /// All proposals to manage the canisters of service nervous systems (SNS), including upgrading
+    /// relevant canisters and managing SNS framework canister WASMs through SNS-W.
     ServiceNervousSystemManagement = 18,
 }
 impl Topic {
@@ -4690,17 +4685,33 @@ impl ProposalRewardStatus {
     }
 }
 
-/// A Request to list minted node provider rewards
+/// A closed range of dates (i.e. includes both start and end dates)
 #[derive(candid::CandidType, candid::Deserialize, serde::Serialize, comparable::Comparable)]
 #[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq)]
-pub struct ListNodeProviderRewardsRequest {}
+#[derive(Debug, Clone, PartialEq)]
+pub struct DateRangeFilter {
+    /// The start date of the range as seconds since epoch.  When not provided,
+    /// no start date is assumed.
+    pub start_timestamp_seconds: Option<u64>,
+    /// The end date of the range as seconds since epoch.  When not provided, no end date is assumed.
+    pub end_timestamp_seconds: Option<u64>,
+}
+
+/// A Request to list minted node provider rewards.  Rewards are listed in descending order of date
+/// minted, meaning that the latest rewards are always returned first.
+#[derive(candid::CandidType, candid::Deserialize, serde::Serialize, comparable::Comparable)]
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Debug, Clone, PartialEq)]
+pub struct ListNodeProviderRewardsRequest {
+    /// Filter for the dates of the rewards
+    pub date_filter: Option<DateRangeFilter>,
+}
 
 /// A Response to list minted node provider rewards.
 /// Includes optional paging information to get next set of results.
 #[derive(candid::CandidType, candid::Deserialize, serde::Serialize, comparable::Comparable)]
 #[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct ListNodeProviderRewardsResponse {
     /// The list of minted node provider rewards
     pub rewards: Vec<MonthlyNodeProviderRewards>,