+++ rs/nns/governance/canister/canister.rs
@@ -33,11 +33,13 @@ use ic_nns_common::{
 use ic_nns_constants::LEDGER_CANISTER_ID;
 use ic_nns_governance::{
     decoder_config, encode_metrics,
-    governance::{Environment, Governance, HeapGrowthPotential, TimeWarp},
+    governance::{Environment, Governance, HeapGrowthPotential, TimeWarp as GovTimeWarp},
     neuron_data_validation::NeuronDataValidationSummary,
     pb::{v1 as gov_pb, v1::Governance as InternalGovernanceProto},
     storage::{grow_upgrades_memory_to, validate_stable_storage, with_upgrades_memory},
 };
+#[cfg(feature = "test")]
+use ic_nns_governance_api::test_api::TimeWarp;
 use ic_nns_governance_api::{
     bitcoin::{BitcoinNetwork, BitcoinSetConfigProposal},
     pb::v1::{
@@ -55,10 +57,9 @@ use ic_nns_governance_api::{
         ListNodeProviderRewardsResponse, ListNodeProvidersResponse, ListProposalInfo,
         ListProposalInfoResponse, ManageNeuronCommandRequest, ManageNeuronRequest,
         ManageNeuronResponse, MonthlyNodeProviderRewards, NetworkEconomics, Neuron, NeuronInfo,
-        NodeProvider, Proposal, ProposalInfo, RestoreAgingSummary, RewardEvent, RewardNodeProvider,
-        RewardNodeProviders, SettleCommunityFundParticipation,
-        SettleNeuronsFundParticipationRequest, SettleNeuronsFundParticipationResponse,
-        UpdateNodeProvider, Vote,
+        NodeProvider, Proposal, ProposalInfo, RestoreAgingSummary, RewardEvent,
+        SettleCommunityFundParticipation, SettleNeuronsFundParticipationRequest,
+        SettleNeuronsFundParticipationResponse, UpdateNodeProvider, Vote,
     },
     subnet_rental::{SubnetRentalProposalPayload, SubnetRentalRequest},
 };
@@ -159,7 +160,7 @@ fn set_governance(gov: Governance) {
 
 struct CanisterEnv {
     rng: ChaCha20Rng,
-    time_warp: TimeWarp,
+    time_warp: GovTimeWarp,
 }
 
 impl CanisterEnv {
@@ -185,7 +186,7 @@ impl CanisterEnv {
                 ChaCha20Rng::from_seed(seed)
             },
 
-            time_warp: TimeWarp { delta_s: 0 },
+            time_warp: GovTimeWarp { delta_s: 0 },
         }
     }
 }
@@ -201,7 +202,7 @@ impl Environment for CanisterEnv {
         )
     }
 
-    fn set_time_warp(&mut self, new_time_warp: TimeWarp) {
+    fn set_time_warp(&mut self, new_time_warp: GovTimeWarp) {
         self.time_warp = new_time_warp;
     }
 
@@ -398,7 +399,7 @@ fn set_time_warp() {
 
 #[cfg(feature = "test")]
 fn set_time_warp_(new_time_warp: TimeWarp) {
-    governance_mut().set_time_warp(new_time_warp);
+    governance_mut().set_time_warp(GovTimeWarp::from(new_time_warp));
 }
 
 /// DEPRECATED: Use manage_neuron directly instead.
@@ -724,17 +725,10 @@ fn get_monthly_node_provider_rewards() {
 }
 
 #[candid_method(update, rename = "get_monthly_node_provider_rewards")]
-async fn get_monthly_node_provider_rewards_() -> Result<RewardNodeProviders, GovernanceError> {
+async fn get_monthly_node_provider_rewards_() -> Result<MonthlyNodeProviderRewards, GovernanceError>
+{
     let rewards = governance_mut().get_monthly_node_provider_rewards().await?;
-    let rewards = rewards
-        .rewards
-        .into_iter()
-        .map(RewardNodeProvider::from)
-        .collect();
-    Ok(RewardNodeProviders {
-        rewards,
-        use_registry_derived_rewards: Some(true),
-    })
+    Ok(MonthlyNodeProviderRewards::from(rewards))
 }
 
 #[export_name = "canister_query list_node_provider_rewards"]
@@ -745,16 +739,15 @@ fn list_node_provider_rewards() {
 
 #[candid_method(query, rename = "list_node_provider_rewards")]
 fn list_node_provider_rewards_(
-    _req: ListNodeProviderRewardsRequest,
+    req: ListNodeProviderRewardsRequest,
 ) -> ListNodeProviderRewardsResponse {
-    let rewards = governance().list_node_provider_rewards(5);
+    let rewards = governance()
+        .list_node_provider_rewards(req.date_filter.map(|d| d.into()))
+        .into_iter()
+        .map(MonthlyNodeProviderRewards::from)
+        .collect();
 
-    ListNodeProviderRewardsResponse {
-        rewards: rewards
-            .into_iter()
-            .map(MonthlyNodeProviderRewards::from)
-            .collect(),
-    }
+    ListNodeProviderRewardsResponse { rewards }
 }
 
 #[export_name = "canister_query list_known_neurons"]
@@ -1277,7 +1270,7 @@ fn test_set_time_warp() {
     let mut environment = CanisterEnv::new();
 
     let start = environment.now();
-    environment.set_time_warp(TimeWarp { delta_s: 1_000 });
+    environment.set_time_warp(GovTimeWarp { delta_s: 1_000 });
     let delta_s = environment.now() - start;
 
     assert!(delta_s >= 1000, "delta_s = {}", delta_s);